+++ Dyalect/Parser/Model/NodeType.cs
@@ -18,6 +18,7 @@
         Binding,
         If,
         While,
+        For,
         Break,
         Continue,
         Return,