+++ Dyalect/Parser/Parser.cs
@@ -40,7 +40,7 @@ namespace Dyalect.Parser
 	public const int _curlyRightToken = 28;
 	public const int _squareLeftToken = 29;
 	public const int _squareRightToken = 30;
-	public const int maxT = 68;
+	public const int maxT = 69;
 
 
 
@@ -74,7 +74,7 @@ namespace Dyalect.Parser
 		} else if (la.kind == 5) {
 			Get();
 			val = ParseChar(); 
-		} else SynErr(69);
+		} else SynErr(70);
 	}
 
 	void StandardOperators() {
@@ -155,7 +155,7 @@ namespace Dyalect.Parser
 			Get();
 			break;
 		}
-		default: SynErr(70); break;
+		default: SynErr(71); break;
 		}
 	}
 
@@ -164,7 +164,7 @@ namespace Dyalect.Parser
 			Get();
 		} else if (StartOf(1)) {
 			StandardOperators();
-		} else SynErr(71);
+		} else SynErr(72);
 	}
 
 	void Qualident(out string s1, out string s2, out string s3) {
@@ -244,7 +244,7 @@ namespace Dyalect.Parser
 			Separator();
 			break;
 		}
-		case 1: case 2: case 3: case 4: case 5: case 25: case 29: case 32: case 38: case 39: case 49: case 65: case 66: case 67: {
+		case 1: case 2: case 3: case 4: case 5: case 25: case 29: case 32: case 38: case 39: case 49: case 66: case 67: case 68: {
 			SimpleExpr(out node);
 			Separator();
 			break;
@@ -272,7 +272,7 @@ namespace Dyalect.Parser
 			If(out node);
 			break;
 		}
-		case 15: {
+		case 14: case 15: {
 			Loops(out node);
 			break;
 		}
@@ -280,7 +280,7 @@ namespace Dyalect.Parser
 			Function(out node);
 			break;
 		}
-		default: SynErr(72); break;
+		default: SynErr(73); break;
 		}
 	}
 
@@ -289,7 +289,7 @@ namespace Dyalect.Parser
 			Get();
 		} else if (la.kind == 8) {
 			Get();
-		} else SynErr(73);
+		} else SynErr(74);
 		var bin = new DBinding(t) { Constant = t.val == "const" }; 
 		Expect(1);
 		bin.Name = t.val; 
@@ -307,7 +307,7 @@ namespace Dyalect.Parser
 			FunctionExpr(out node);
 		} else if (StartOf(3)) {
 			Assignment(out node);
-		} else SynErr(74);
+		} else SynErr(75);
 	}
 
 	void ControlFlow(out DNode node) {
@@ -318,7 +318,7 @@ namespace Dyalect.Parser
 			Continue(out node);
 		} else if (la.kind == 10) {
 			Return(out node);
-		} else SynErr(75);
+		} else SynErr(76);
 	}
 
 	void Block(out DNode node) {
@@ -351,14 +351,18 @@ namespace Dyalect.Parser
 			} else if (la.kind == 13) {
 				If(out node);
 				@if.False = node; 
-			} else SynErr(76);
+			} else SynErr(77);
 		}
 		node = @if; 
 	}
 
 	void Loops(out DNode node) {
 		node = null; 
-		While(out node);
+		if (la.kind == 15) {
+			While(out node);
+		} else if (la.kind == 14) {
+			For(out node);
+		} else SynErr(78);
 	}
 
 	void Function(out DNode node) {
@@ -388,7 +392,7 @@ namespace Dyalect.Parser
 			Get();
 			Expr(out node);
 			Separator();
-		} else SynErr(77);
+		} else SynErr(79);
 		f.Body = node;
 		node = f;
 		
@@ -420,9 +424,9 @@ namespace Dyalect.Parser
 			If(out node);
 		} else if (la.kind == 27) {
 			Block(out node);
-		} else if (la.kind == 15) {
+		} else if (la.kind == 14 || la.kind == 15) {
 			Loops(out node);
-		} else SynErr(78);
+		} else SynErr(80);
 	}
 
 	void FunctionArgument(DFunctionDeclaration node) {
@@ -440,6 +444,26 @@ namespace Dyalect.Parser
 		@while.Body = node;
 		node = @while;
 		
+	}
+
+	void For(out DNode node) {
+		node = null; 
+		Expect(14);
+		var @for = new DFor(t); 
+		Name(out node);
+		@for.Variable = (DName)node; 
+		Expect(53);
+		Expr(out node);
+		@for.Target = node; 
+		Block(out node);
+		@for.Body = node;
+		node = @for;
+		
+	}
+
+	void Name(out DNode node) {
+		Expect(1);
+		node = new DName(t) { Value = t.val }; 
 	}
 
 	void Break(out DNode node) {
@@ -473,7 +497,7 @@ namespace Dyalect.Parser
 			FunctionArgument(f);
 		} else if (la.kind == 25) {
 			FunctionArguments(f);
-		} else SynErr(79);
+		} else SynErr(81);
 		Expect(19);
 		Expr(out var exp);
 		f.Body = exp; 
@@ -491,52 +515,52 @@ namespace Dyalect.Parser
 				Get();
 				break;
 			}
-			case 53: {
+			case 54: {
 				Get();
 				op = BinaryOperator.Add; 
 				break;
 			}
-			case 54: {
+			case 55: {
 				Get();
 				op = BinaryOperator.Sub; 
 				break;
 			}
-			case 55: {
+			case 56: {
 				Get();
 				op = BinaryOperator.Mul; 
 				break;
 			}
-			case 56: {
+			case 57: {
 				Get();
 				op = BinaryOperator.Div; 
 				break;
 			}
-			case 57: {
+			case 58: {
 				Get();
 				op = BinaryOperator.Rem; 
 				break;
 			}
-			case 58: {
+			case 59: {
 				Get();
 				op = BinaryOperator.And; 
 				break;
 			}
-			case 59: {
+			case 60: {
 				Get();
 				op = BinaryOperator.Or; 
 				break;
 			}
-			case 60: {
+			case 61: {
 				Get();
 				op = BinaryOperator.Xor; 
 				break;
 			}
-			case 61: {
+			case 62: {
 				Get();
 				op = BinaryOperator.ShiftLeft; 
 				break;
 			}
-			case 62: {
+			case 63: {
 				Get();
 				op = BinaryOperator.ShiftRight; 
 				break;
@@ -552,7 +576,7 @@ namespace Dyalect.Parser
 
 	void Or(out DNode node) {
 		And(out node);
-		while (la.kind == 63) {
+		while (la.kind == 64) {
 			Get();
 			var ot = t; 
 			And(out DNode exp);
@@ -562,7 +586,7 @@ namespace Dyalect.Parser
 
 	void And(out DNode node) {
 		Eq(out node);
-		while (la.kind == 64) {
+		while (la.kind == 65) {
 			Get();
 			var ot = t; 
 			Eq(out DNode exp);
@@ -787,9 +811,9 @@ namespace Dyalect.Parser
 			Float(out node);
 		} else if (la.kind == 4 || la.kind == 5) {
 			String(out node);
-		} else if (la.kind == 66 || la.kind == 67) {
+		} else if (la.kind == 67 || la.kind == 68) {
 			Bool(out node);
-		} else if (la.kind == 65) {
+		} else if (la.kind == 66) {
 			Nil(out node);
 		} else if (IsTuple()) {
 			Tuple(out node);
@@ -797,12 +821,7 @@ namespace Dyalect.Parser
 			Array(out node);
 		} else if (la.kind == 25) {
 			Group(out node);
-		} else SynErr(80);
-	}
-
-	void Name(out DNode node) {
-		Expect(1);
-		node = new DName(t) { Value = t.val }; 
+		} else SynErr(82);
 	}
 
 	void Integer(out DNode node) {
@@ -821,16 +840,16 @@ namespace Dyalect.Parser
 	}
 
 	void Bool(out DNode node) {
-		if (la.kind == 66) {
+		if (la.kind == 67) {
 			Get();
-		} else if (la.kind == 67) {
+		} else if (la.kind == 68) {
 			Get();
-		} else SynErr(81);
+		} else SynErr(83);
 		node = new DBooleanLiteral(t) { Value = t.val == "true" }; 
 	}
 
 	void Nil(out DNode node) {
-		Expect(65);
+		Expect(66);
 		node = new DNilLiteral(t); 
 	}
 
@@ -881,14 +900,14 @@ namespace Dyalect.Parser
 			} else if (la.kind == 4 || la.kind == 5) {
 				AnyString(out var str);
 				name = str; 
-			} else SynErr(82);
+			} else SynErr(84);
 			Expect(23);
 			var tag = new DLabelLiteral(t) { Label = name }; 
 			Expr(out node);
 			tag.Expression = node; node = tag; 
 		} else if (StartOf(5)) {
 			Expr(out node);
-		} else SynErr(83);
+		} else SynErr(85);
 	}
 
 	void Dyalect() {
@@ -913,15 +932,15 @@ namespace Dyalect.Parser
         }
 
         static readonly bool[,] set = {
-		{_T,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x},
-		{_x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x},
-		{_x,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _T,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x},
-		{_x,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_x, _x,_T,_x,_x, _T,_x,_x,_x, _x,_x,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_T,_T, _x,_x},
-		{_x,_T,_T,_T, _T,_T,_x,_T, _T,_T,_T,_T, _T,_T,_x,_T, _x,_x,_T,_x, _x,_x,_x,_x, _x,_T,_x,_T, _x,_T,_x,_x, _T,_x,_x,_x, _x,_x,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_T,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_T,_T, _x,_x},
-		{_x,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_T,_x,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_T, _x,_T,_x,_x, _T,_x,_x,_x, _x,_x,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_T,_T, _x,_x},
-		{_x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _T,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_x, _x,_x,_x,_x, _x,_x},
-		{_x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _T,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x},
-		{_x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _T,_x,_x,_x, _x,_x,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x}
+		{_T,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x},
+		{_x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x},
+		{_x,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _T,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x},
+		{_x,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_x, _x,_T,_x,_x, _T,_x,_x,_x, _x,_x,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_T,_T, _T,_x,_x},
+		{_x,_T,_T,_T, _T,_T,_x,_T, _T,_T,_T,_T, _T,_T,_T,_T, _x,_x,_T,_x, _x,_x,_x,_x, _x,_T,_x,_T, _x,_T,_x,_x, _T,_x,_x,_x, _x,_x,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_T,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_T,_T, _T,_x,_x},
+		{_x,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_T,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_T, _x,_T,_x,_x, _T,_x,_x,_x, _x,_x,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_T,_T, _T,_x,_x},
+		{_x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _T,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_T,_T, _T,_T,_T,_T, _T,_T,_T,_T, _x,_x,_x,_x, _x,_x,_x},
+		{_x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _T,_T,_T,_T, _T,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x},
+		{_x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _T,_x,_x,_x, _x,_x,_T,_T, _x,_x,_x,_x, _x,_x,_x,_x, _x,_T,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x,_x, _x,_x,_x}
 
         };
 
@@ -984,37 +1003,39 @@ namespace Dyalect.Parser
 			case 50: s = "\"import\" expected"; break;
 			case 51: s = "\"...\" expected"; break;
 			case 52: s = "\"else\" expected"; break;
-			case 53: s = "\"+=\" expected"; break;
-			case 54: s = "\"-=\" expected"; break;
-			case 55: s = "\"*=\" expected"; break;
-			case 56: s = "\"/=\" expected"; break;
-			case 57: s = "\"%=\" expected"; break;
-			case 58: s = "\"&=\" expected"; break;
-			case 59: s = "\"|=\" expected"; break;
-			case 60: s = "\"^=\" expected"; break;
-			case 61: s = "\"<<=\" expected"; break;
-			case 62: s = "\">>=\" expected"; break;
-			case 63: s = "\"||\" expected"; break;
-			case 64: s = "\"&&\" expected"; break;
-			case 65: s = "\"nil\" expected"; break;
-			case 66: s = "\"true\" expected"; break;
-			case 67: s = "\"false\" expected"; break;
-			case 68: s = "??? expected"; break;
-			case 69: s = "invalid AnyString"; break;
-			case 70: s = "invalid StandardOperators"; break;
-			case 71: s = "invalid FunctionName"; break;
-			case 72: s = "invalid Statement"; break;
-			case 73: s = "invalid Binding"; break;
-			case 74: s = "invalid SimpleExpr"; break;
-			case 75: s = "invalid ControlFlow"; break;
-			case 76: s = "invalid If"; break;
-			case 77: s = "invalid Function"; break;
-			case 78: s = "invalid Expr"; break;
-			case 79: s = "invalid FunctionExpr"; break;
-			case 80: s = "invalid Literal"; break;
-			case 81: s = "invalid Bool"; break;
-			case 82: s = "invalid TupleElement"; break;
-			case 83: s = "invalid TupleElement"; break;
+			case 53: s = "\"in\" expected"; break;
+			case 54: s = "\"+=\" expected"; break;
+			case 55: s = "\"-=\" expected"; break;
+			case 56: s = "\"*=\" expected"; break;
+			case 57: s = "\"/=\" expected"; break;
+			case 58: s = "\"%=\" expected"; break;
+			case 59: s = "\"&=\" expected"; break;
+			case 60: s = "\"|=\" expected"; break;
+			case 61: s = "\"^=\" expected"; break;
+			case 62: s = "\"<<=\" expected"; break;
+			case 63: s = "\">>=\" expected"; break;
+			case 64: s = "\"||\" expected"; break;
+			case 65: s = "\"&&\" expected"; break;
+			case 66: s = "\"nil\" expected"; break;
+			case 67: s = "\"true\" expected"; break;
+			case 68: s = "\"false\" expected"; break;
+			case 69: s = "??? expected"; break;
+			case 70: s = "invalid AnyString"; break;
+			case 71: s = "invalid StandardOperators"; break;
+			case 72: s = "invalid FunctionName"; break;
+			case 73: s = "invalid Statement"; break;
+			case 74: s = "invalid Binding"; break;
+			case 75: s = "invalid SimpleExpr"; break;
+			case 76: s = "invalid ControlFlow"; break;
+			case 77: s = "invalid If"; break;
+			case 78: s = "invalid Loops"; break;
+			case 79: s = "invalid Function"; break;
+			case 80: s = "invalid Expr"; break;
+			case 81: s = "invalid FunctionExpr"; break;
+			case 82: s = "invalid Literal"; break;
+			case 83: s = "invalid Bool"; break;
+			case 84: s = "invalid TupleElement"; break;
+			case 85: s = "invalid TupleElement"; break;
 
                 default:
                     s = "unknown " + n;