+++ Dyalect/Runtime/Types/DyString.cs
@@ -116,7 +116,20 @@ namespace Dyalect.Runtime.Types
 
         protected override DyString ToStringOp(DyObject arg, ExecutionContext ctx) => StringUtil.Escape(arg.GetString());
 
-        private DyObject GetIterator(DyObject obj, ExecutionContext ctx) => new DyStringIterator((DyString)obj);
+        private DyFunction GetIterator(DyObject obj, ExecutionContext ctx)
+        {
+            var idx = 0;
+            var str = (DyString)obj;
+
+            return DyFunction.Create(
+                () =>
+                {
+                    if (idx < str.Value.Length)
+                        return DyTuple.Create(DyBool.True, new DyString(str.Value[idx++].ToString()));
+
+                    return DyTuple.Create(DyBool.False, DyNil.Instance);
+                }, "iterator");
+        }
 
         protected override DyFunction GetTrait(string name, ExecutionContext ctx)
         {
@@ -126,60 +139,5 @@ namespace Dyalect.Runtime.Types
             return null;
         }
         #endregion
-    }
-
-    internal sealed class DyStringIterator : DyObject
-    {
-        internal DyString String { get; }
-
-        internal int Index { get; set; } = -1;
-
-        public DyStringIterator(DyString str) : base(StandardType.StringIterator)
-        {
-            String = str;
-        }
-
-        public override object ToObject() => ((IEnumerable<object>)String.ToObject()).GetEnumerator();
-
-        protected override bool TestEquality(DyObject obj) => String.Equals(((DyStringIterator)obj).String);
-    }
-
-    internal sealed class DyStringIteratorTypeInfo : DyTypeInfo
-    {
-        public static readonly DyTypeInfo Instance = new DyStringIteratorTypeInfo();
-
-        public DyStringIteratorTypeInfo() : base(StandardType.StringIterator)
-        {
-
-        }
-
-        protected override DyObject MoveNextOp(DyObject obj, ExecutionContext ctx)
-        {
-            var iter = (DyStringIterator)obj;
-            iter.Index++;
-            return iter.Index < iter.String.Value.Length ? DyBool.True : DyBool.False;
-        }
-
-        protected override DyObject GetCurrentOp(DyObject obj, ExecutionContext ctx)
-        {
-            var iter = (DyStringIterator)obj;
-            return new DyString(iter.String.Value[iter.Index].ToString());
-        }
-
-        protected override DyFunction GetTrait(string name, ExecutionContext ctx)
-        {
-            if (name == Traits.NextName)
-                return new DyMemberFunction(name, MoveNextOp);
-
-            if (name == Traits.CurName)
-                return new DyMemberFunction(name, GetCurrentOp);
-
-            return null;
-        }
-
-        public override string TypeName => StandardType.StringIteratorName;
-
-        public override DyObject Create(ExecutionContext ctx, params DyObject[] args) =>
-            throw new NotImplementedException();
     }
 }