+++ Docs/Changes.md
@@ -3,6 +3,8 @@
     invocation code.
  * A bug fixed in standard `toString` function implementation
     that could cause virtual machine to crash.
+ * Fixed compilation logic for `while` loops (previous they
+    didn't create a lexical scope).
 
 # 0.1.0
 Initial release